Added JWT Authentication and admin and users roles authorization

+++ Application.Api/Controllers/ProductsController.cs
@@ -1,3 +1,5 @@
+using Application.Api.Authorization;
+using Application.Api.Data;
 using Application.Api.Models;
 using Application.Api.Services.Products;
 using Microsoft.AspNetCore.Mvc;
@@ -8,18 +10,27 @@ namespace Application.Api.Controllers;
 [Route("[controller]")]
 public class ProductsController : ControllerBase
 {
-    // TODO: Make this Authorized only
     private readonly IProductsService _productService;
+    private readonly ApplicationContext _context;
 
-    public ProductsController(IProductsService productsService)
+    public ProductsController(IProductsService productsService, ApplicationContext context)
     {
         _productService = productsService;
+        _context = context;
     }
 
     [HttpPost]
     [Route("add")]
-    public IActionResult AddProduct(Product product)
+    public IActionResult AddProduct(Product product, Guid userId)
     {
+        // TODO: Make this as attribute
+        var user = _context.Users.First(u => u.Id == userId);
+
+        if (user.Roles.Any(role => role == ERoles.User))
+        {
+            return Unauthorized();
+        }
+        
         _productService.AddProduct(product);
         return Ok();
     }
@@ -47,8 +58,16 @@ public class ProductsController : ControllerBase
 
     [HttpDelete]
     [Route("remove")]
-    public IActionResult RemoveProduct(int id)
+    public IActionResult RemoveProduct(int id, Guid userId)
     {
+        // TODO: Make this as attribute
+        var user = _context.Users.First(u => u.Id == userId);
+
+        if (user.Roles.Any(role => role == ERoles.User))
+        {
+            return Unauthorized();
+        }
+        
         _productService.RemoveProduct(id);
         return Ok();
     }
